backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0

# Data
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.26.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Scheduler
apscheduler>=3.10.4

# News & Intelligence (Phase 2 & 5)
feedparser>=6.0.0
youtube-transcript-api>=0.6.2
google-genai>=0.1.0
python-multipart>=0.0.9
